#15: Refactored TxtStandard
